Extract change set parsing logic into dedicated SegmentChangeSetsParser class
